Rss feed fix (#293)
* Update social icons in footer

Add to the site footer an icon linking to the YouTube channel as per issue #2239. Change the email entry to direct to info@interlisp.org instead of the Interlisp Users mailing list, which addresses issue #1426.

Signed-off-by: Paolo Amoroso <info@paoloamoroso.com>

* Fixes rss feed.  Changes to _index.md are now included.

The code was including all the pages within a section but not the
section index, _index.md.  This was fixed by specifically adding
the section index page:

```
  {{- $pages = $pages | union (slice .) }}
```

and then adding the pages within the section:

```
{{- $pages = $pages | union (slice .) }}
```

In addition, the pages are ordered via Publish Date before the
resulting feed is build.

---------

Signed-off-by: Paolo Amoroso <info@paoloamoroso.com>
Co-authored-by: Paolo Amoroso <info@paoloamoroso.com>